set up logic for categories implementation

+++ src/main/java/Services/CategoryServiceImpl.java
@@ -5,14 +5,20 @@
 package Services;
 
 import Data.Category;
+import Data.User;
 import Exceptions.ActionDeniedException;
+import Exceptions.ResourceAccessDeniedException;
 import Exceptions.ResourceAlreadyExistsException;
 import Exceptions.ResourceNotFoundException;
 import Exceptions.ValidationException;
 import Mappers.CategoryMapper;
+import Mappers.impl.CategoryMapperImpl;
 import Models.CreateCategoryRequest;
 import Models.UpdateCategoryRequest;
 import Repositories.CategoryRepository;
+import Repositories.UserRepository;
+import Repositories.impl.CategoryRepositoryHibernete;
+import Repositories.impl.UserRepositoryHibernete;
 import Validators.RequestValidator;
 import java.util.List;
 
@@ -23,50 +29,72 @@ import java.util.List;
 public class CategoryServiceImpl implements CategoryService {
 
     private final CategoryRepository categoryRepository;
+    private final UserRepository userRepository;
+    private final CategoryMapper categoryMapper;
 
-    public CategoryServiceImpl(CategoryRepository categoryRepository) {
-        this.categoryRepository = categoryRepository;
+    public CategoryServiceImpl() {
+        this.categoryRepository = new CategoryRepositoryHibernete();
+        this.userRepository = new UserRepositoryHibernete();
+        this.categoryMapper = new CategoryMapperImpl();
     }
 
     @Override
-    public Category create(CreateCategoryRequest request)
+    public Category create(CreateCategoryRequest request, int userId)
             throws ValidationException, ResourceAlreadyExistsException {
 
         RequestValidator.validate(request);
 
+        /*
         if (categoryRepository.findByName(request.getName()) != null) {
+            throw new ResourceAlreadyExistsException();
+        }*/
+        if (categoryRepository.existsByNameAndUser(request.getName(), userId)) {
             throw new ResourceAlreadyExistsException();
         }
 
-        Category category = CategoryMapper.toEntity(request);
+        User user = userRepository.findById(userId);
+
+        Category category = categoryMapper.toEntity(request, user);
+
         categoryRepository.save(category);
         return category;
     }
 
     @Override
-    public Category update(UpdateCategoryRequest request)
+    public Category update(UpdateCategoryRequest request, int userId)
             throws ValidationException, ResourceNotFoundException {
 
         RequestValidator.validate(request);
 
+        /*
         Category category = categoryRepository.findById(request.getId());
+         */
+        Category category
+                = categoryRepository.findByIdAndUser(request.getId(), userId);
+
         if (category == null) {
             throw new ResourceNotFoundException();
         }
 
-        CategoryMapper.updateEntity(category, request);
+        categoryMapper.updateEntity(category, request);
         categoryRepository.update(category);
         return category;
     }
 
     @Override
-    public List<Category> getAll() {
-        return categoryRepository.findAll();
+    public List<Category> getAll(int userId) {
+        return categoryRepository.findAll(userId);
     }
 
     @Override
-    public void delete(int id) throws ActionDeniedException {
-        // optional: prevent deletion if used by todos
+    public void delete(int id, int userId) {
+
+        Category category
+                = categoryRepository.findByIdAndUser(id, userId);
+
+        if (category == null) {
+            throw new ResourceAccessDeniedException();
+        }
         categoryRepository.delete(id);
     }
 }
